Separate SpotPicker from guiutils.h
